Remove long unused JMX collector interface types.
